+++ cmd/internal/cli/actions.go
@@ -28,11 +28,13 @@ import (
 	"github.com/apptainer/apptainer/internal/pkg/util/env"
 	"github.com/apptainer/apptainer/internal/pkg/util/uri"
 	"github.com/apptainer/apptainer/pkg/sylog"
+	"github.com/apptainer/apptainer/pkg/util/fs/lock"
 	"github.com/spf13/cobra"
 )
 
 const (
-	defaultPath = "/bin:/usr/bin:/sbin:/usr/sbin:/usr/local/bin:/usr/local/sbin"
+	defaultPath           = "/bin:/usr/bin:/sbin:/usr/sbin:/usr/local/bin:/usr/local/sbin"
+	shareNSInstancePrefix = "sharens_instance"
 )
 
 func getCacheHandle(cfg cache.Config) *cache.Handle {
@@ -175,8 +177,14 @@ var ExecCmd = &cobra.Command{
 	PreRun:                actionPreRun,
 	Run: func(cmd *cobra.Command, args []string) {
 		a := append([]string{"/.singularity.d/actions/exec"}, args[1:]...)
-		if err := launchContainer(cmd, args[0], a, ""); err != nil {
-			sylog.Fatalf("%s", err)
+		if shareNS {
+			if err := shareNSLaunch(cmd, args[0], a); err != nil {
+				sylog.Fatalf("%s", err)
+			}
+		} else {
+			if err := launchContainer(cmd, args[0], a, "", -1); err != nil {
+				sylog.Fatalf("%s", err)
+			}
 		}
 	},
 
@@ -198,8 +206,14 @@ var ShellCmd = &cobra.Command{
 		}
 
 		a := []string{"/.singularity.d/actions/shell"}
-		if err := launchContainer(cmd, args[0], a, ""); err != nil {
-			sylog.Fatalf("%s", err)
+		if shareNS {
+			if err := shareNSLaunch(cmd, args[0], a); err != nil {
+				sylog.Fatalf("%s", err)
+			}
+		} else {
+			if err := launchContainer(cmd, args[0], a, "", -1); err != nil {
+				sylog.Fatalf("%s", err)
+			}
 		}
 	},
 
@@ -217,8 +231,14 @@ var RunCmd = &cobra.Command{
 	PreRun:                actionPreRun,
 	Run: func(cmd *cobra.Command, args []string) {
 		a := append([]string{"/.singularity.d/actions/run"}, args[1:]...)
-		if err := launchContainer(cmd, args[0], a, ""); err != nil {
-			sylog.Fatalf("%s", err)
+		if shareNS {
+			if err := shareNSLaunch(cmd, args[0], a); err != nil {
+				sylog.Fatalf("%s", err)
+			}
+		} else {
+			if err := launchContainer(cmd, args[0], a, "", -1); err != nil {
+				sylog.Fatalf("%s", err)
+			}
 		}
 	},
 
@@ -236,8 +256,14 @@ var TestCmd = &cobra.Command{
 	PreRun:                actionPreRun,
 	Run: func(cmd *cobra.Command, args []string) {
 		a := append([]string{"/.singularity.d/actions/test"}, args[1:]...)
-		if err := launchContainer(cmd, args[0], a, ""); err != nil {
-			sylog.Fatalf("%s", err)
+		if shareNS {
+			if err := shareNSLaunch(cmd, args[0], a); err != nil {
+				sylog.Fatalf("%s", err)
+			}
+		} else {
+			if err := launchContainer(cmd, args[0], a, "", -1); err != nil {
+				sylog.Fatalf("%s", err)
+			}
 		}
 	},
 
@@ -247,7 +273,7 @@ var TestCmd = &cobra.Command{
 	Example: docs.RunTestExample,
 }
 
-func launchContainer(cmd *cobra.Command, image string, args []string, instanceName string) error {
+func launchContainer(cmd *cobra.Command, image string, args []string, instanceName string, fd int) error {
 	ns := launch.Namespaces{
 		User:  userNamespace,
 		UTS:   utsNamespace,
@@ -322,6 +348,8 @@ func launchContainer(cmd *cobra.Command, image string, args []string, instanceNa
 		launch.OptUseBuildConfig(useBuildConfig),
 		launch.OptTmpDir(tmpDir),
 		launch.OptUnderlay(underlay),
+		launch.OptShareNSMode(shareNS),
+		launch.OptShareNSFd(fd),
 	}
 
 	l, err := launch.NewLauncher(opts...)
@@ -330,4 +358,35 @@ func launchContainer(cmd *cobra.Command, image string, args []string, instanceNa
 	}
 
 	return l.Exec(cmd.Context(), image, args, instanceName)
+}
+
+func shareNSLaunch(cmd *cobra.Command, image string, args []string) error {
+	ppid := os.Getppid()
+	procDir := fmt.Sprintf("/proc/%d", ppid)
+	instanceName := fmt.Sprintf("%s_%d", shareNSInstancePrefix, ppid)
+
+	// try locking the ppid proc folder
+	fd, accquired, err := lock.TryExclusive(procDir)
+	if err != nil {
+		return err
+	}
+
+	if accquired {
+		// first process
+		if err := launchContainer(cmd, image, args, instanceName, fd); err != nil {
+			return err
+		}
+	} else {
+		// other process, this will block the process
+		fd, err := lock.Exclusive(procDir)
+		if err != nil {
+			return err
+		} else if err = lock.Release(fd); err != nil {
+			return err
+		}
+		if err := launchContainer(cmd, fmt.Sprintf("instance://%s", instanceName), args, "", -1); err != nil {
+			return err
+		}
+	}
+	return nil
 }